refactor: Move Zoom webhook handling to platform-specific structure
Move Zoom webhook handler and validator from generic webhook package
to zoom/webhook subpackage for better organization and scalability.

This change:
- Improves cohesion by keeping all Zoom-related code together
- Scales better for future platforms (Teams, Google Meet, etc.)
- Maintains clear separation between platform-agnostic webhook
  registry and platform-specific webhook implementations
